requirements.txt
aiogram>=3.6
aiosqlite>=0.19
APScheduler>=3.10
python-dotenv>=1.0
aiohttp>=3.9
telethon>=1.36